Fix lint and list compatibility
